+++ stocc_random_number_generator/stoc1.cpp
@@ -44,7 +44,7 @@ double LnFac(int32 n) {
       // make table of ln(n!)
       double sum = fac_table[0] = 0.;
       for (int i=1; i<FAK_LEN; i++) {
-        sum += log(i);
+        sum += log(static_cast<double>(i));
         fac_table[i] = sum;}
       initialized = 1;}
     return fac_table[n];}
